+++ backend/src/main/java/com/celuveat/celuveat/admin/presentation/AdminAuthController.java
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/admin")
 @RequiredArgsConstructor
-public class AdminController {
+public class AdminAuthController {
 
     private final AdminService adminService;
