+++ backend/src/test/java/com/celuveat/celuveat/acceptance/admin/auth/AdminAuthAcceptanceTest.java
@@ -1,6 +1,6 @@
-package com.celuveat.celuveat.acceptance.admin;
+package com.celuveat.celuveat.acceptance.admin.auth;
 
-import static com.celuveat.celuveat.acceptance.admin.AdminAcceptanceSteps.관리자_로그인_요청;
+import static com.celuveat.celuveat.acceptance.admin.auth.AdminAuthAcceptanceSteps.관리자_로그인_요청;
 import static com.celuveat.celuveat.acceptance.common.CommonAcceptanceSteps.SESSION_ID_IN_COOKIE;
 import static com.celuveat.celuveat.acceptance.common.CommonAcceptanceSteps.값이_존재한다;
 import static com.celuveat.celuveat.acceptance.common.CommonAcceptanceSteps.발생한_예외를_검증한다;
@@ -15,11 +15,14 @@ import com.celuveat.celuveat.acceptance.common.AcceptanceTest;
 import com.celuveat.celuveat.admin.domain.Admin;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 
 @DisplayName("Admin 인수테스트")
 @SuppressWarnings("NonAsciiCharacters")
-class AdminAcceptanceTest extends AcceptanceTest {
+@DisplayNameGeneration(ReplaceUnderscores.class)
+class AdminAuthAcceptanceTest extends AcceptanceTest {
 
     @BeforeEach
     void setUp() {
